Split Legend and Metrics, into small files
