+++ src/components/Providers.tsx
@@ -7,7 +7,7 @@ import { MetricsContext } from "../contexts/MetricContext"
 import { ChartType, getDefaultOptions, Options, OptionsContext } from "../contexts/OptionsContext"
 import { PathContext } from "../contexts/PathContext"
 import { SearchContext } from "../contexts/SearchContext"
-import { AuthorshipType, createMetricData as createMetricsData, MetricsData, MetricType } from "../metrics"
+import { AuthorshipType, createMetricData as createMetricsData, MetricsData, MetricType } from "../metrics/metrics"
 
 interface ProvidersProps {
   children: React.ReactNode
