+++ src/components/Tooltip.tsx
@@ -5,7 +5,7 @@ import { HydratedGitBlobObject } from "~/analyzer/model"
 import { useMetrics } from "../contexts/MetricContext"
 import { useOptions } from "../contexts/OptionsContext"
 import { useCSSVar } from "../hooks"
-import { AuthorshipType, MetricType } from "../metrics"
+import { AuthorshipType, MetricType } from "../metrics/metrics"
 import { dateFormatRelative } from "../util"
 import { Spacer } from "./Spacer"
 import { Box, BoxSubTitle, LegendDot } from "./util"
