+++ src/metrics/metrics.ts
@@ -1,8 +1,15 @@
 import distinctColors from "distinct-colors"
 import { AnalyzerData, HydratedGitBlobObject, HydratedGitTreeObject } from "~/analyzer/model"
-import { getColorFromExtension } from "./extension-color"
-import { LegendType } from "./components/Legend"
-import { dateFormatShort } from "./util"
+import { LegendType } from "../components/legend/Legend"
+import { setExtensionColor } from "./fileExtension"
+import { setDominanceColor } from "./singleAuthor"
+import { CommitAmountTranslater } from "./mostCommits"
+import { getLastChangedIndex, lastChangedGroupings } from "./lastChanged"
+import { setDominantAuthorColor } from "./topContributer"
+import { TruckFactorTranslater } from "./truckFactor"
+import { GradLegendData } from "~/components/legend/GradiantLegend"
+import { SegmentLegendData } from "~/components/legend/SegmentLegend"
+import { PointInfo, PointLegendData } from "~/components/legend/PointLegend"
 
 export type MetricsData = [
   Record<AuthorshipType, Map<MetricType, MetricCache>>,
@@ -77,30 +84,6 @@ export function getMetricLegendType(metric: MetricType) : LegendType {
   }
 }
 
-export class PointInfo {
-  constructor(public readonly color: string, public weight: number) {}
-
-  add(value: number) {
-    this.weight += value
-  }
-}
-
-export type PointLegendData = Map<string, PointInfo>
-export type GradLegendData = [
-  minValue: string,
-  maxValue: string,
-  minValueAltFormat: string | undefined,
-  maxValueAltFormat: string | undefined,
-  minColor: string,
-  maxColor: string
-]
-export type SegmentLegendData = [
-  steps: number,
-  textGenerator: (n: number) => string,
-  colorGenerator: (n: number) => string,
-  offsetStepCalc: (blob: HydratedGitBlobObject) => number
-]
-
 export interface MetricCache {
   legend: PointLegendData | GradLegendData | SegmentLegendData| undefined
   colormap: Map<string, string>
@@ -127,8 +110,8 @@ export function getMetricCalcs(
   authorColors: Map<string, string>,
 ): [metricType: MetricType, func: (blob: HydratedGitBlobObject, cache: MetricCache) => void][] {
   const commit = data.commit
-  const heatmap = new HeatMapTranslater(commit.minNoCommits, commit.maxNoCommits)
-  const truckmap = new TruckFactorTranslater(data.authorsUnion.length)
+  const commitmapper = new CommitAmountTranslater(commit.minNoCommits, commit.maxNoCommits)
+  const truckmapper = new TruckFactorTranslater(data.authorsUnion.length)
 
   return [
     [
@@ -156,11 +139,11 @@ export function getMetricCalcs(
             `${commit.maxNoCommits}`,
             undefined,
             undefined,
-            heatmap.getColor(commit.minNoCommits),
-            heatmap.getColor(commit.maxNoCommits),
+            commitmapper.getColor(commit.minNoCommits),
+            commitmapper.getColor(commit.maxNoCommits),
           ]
         }
-        heatmap.setColor(blob, cache)
+        commitmapper.setColor(blob, cache)
       },
     ],
     [
@@ -198,7 +181,7 @@ export function getMetricCalcs(
             (blob) => Math.floor(Math.log2(Object.entries(blob.unionedAuthors?.HISTORICAL ?? []).length))
           ]
         }
-        truckmap.setColor(blob, cache)
+        truckmapper.setColor(blob, cache)
       },
     ],
   ]
@@ -240,179 +223,5 @@ function setupMetricsCacheRec(
         }
         break
     }
-  }
-}
-
-function setExtensionColor(blob: HydratedGitBlobObject, cache: MetricCache) {
-  const extension = blob.name.substring(blob.name.lastIndexOf(".") + 1)
-  const color = getColorFromExtension(extension)
-  const legend = cache.legend as PointLegendData
-  if (color) {
-    if (legend.has(extension)) {
-      legend.get(extension)?.add(1)
-    } else {
-      legend.set(extension, new PointInfo(color, 1))
-    }
-    cache.colormap.set(blob.path, color)
-  } else {
-    if (!legend.has("Other")) legend.set("Other", new PointInfo("grey", 0))
-    cache.colormap.set(blob.path, "grey")
-  }
-}
-
-function setDominantAuthorColor(
-  authorColors: Map<string, string>,
-  blob: HydratedGitBlobObject,
-  cache: MetricCache,
-  authorshipType: AuthorshipType
-) {
-  const authorUnion = blob.unionedAuthors?.[authorshipType]
-  let sorted: [string, number][]
-  try {
-    if (!authorUnion) throw Error
-    sorted = Object.entries(authorUnion).sort(([k1, v1], [k2, v2]) => {
-      if (v1 === 0 || v2 === 0 || !k1 || !k2) throw Error
-      if (v1 < v2) return 1
-      else if (v1 > v2) return -1
-      else return 0
-    })
-    if (!sorted[0]) throw Error
-  } catch {
-    return
-  }
-
-  const [dom] = sorted[0]
-  const legend = cache.legend as PointLegendData
-  const color = authorColors.get(dom) ?? "grey"
-
-  cache.colormap.set(blob.path, color)
-  blob.dominantAuthor?.set(authorshipType, sorted[0])
-
-  if (legend.has(dom)) {
-    legend.get(dom)?.add(1)
-    return
-  }
-  legend.set(dom, new PointInfo(color, 1))
-}
-
-function setDominanceColor(blob: HydratedGitBlobObject, cache: MetricCache, authorshipType: AuthorshipType) {
-  const dominatedColor = "red"
-  const defaultColor = "hsl(210, 38%, 85%)"
-  const nocreditColor = "teal"
-
-  const authorUnion = blob.unionedAuthors?.[authorshipType] ?? {}
-
-  let creditsum = 0
-  for (const [, val] of Object.entries(authorUnion)) {
-    creditsum += val
-  }
-
-  const legend = cache.legend as PointLegendData
-
-  if (creditsum === 0) {
-    legend.set("No authors", new PointInfo(nocreditColor, 0))
-    cache.colormap.set(blob.path, nocreditColor)
-    return
-  }
-
-  if (!authorUnion) throw Error("No unioned authors found")
-  switch (Object.keys(authorUnion).length) {
-    case 1:
-      legend.set("Single author", new PointInfo(dominatedColor, 2))
-      cache.colormap.set(blob.path, dominatedColor)
-      return
-    default:
-      legend.set("Multiple authors", new PointInfo(defaultColor, 1))
-      cache.colormap.set(blob.path, defaultColor)
-      return
-  }
-}
-
-interface lastChangedGroup {
-  epoch: number
-  color: string
-  text: string
-}
-
-function lastChangedGroupings(newestEpoch: number): lastChangedGroup[] {
-  return [
-    {epoch: 0, text: dateFormatShort(newestEpoch*1000), color: "#cff2ff" },
-    {epoch: 86400, text: "+1d", color: "#c6dbef"},
-    {epoch: 604800, text: "+1w", color: "#9ecae1" },
-    {epoch: 2629743, text: "+1m", color: "#6baed6" },
-    {epoch: 31556926, text: "+1y", color: "#4292c6" },
-    {epoch: 31556926*2, text: "+2y", color: "#2171b5" },
-    {epoch: 31556926*4, text: "+4y", color: "#084594" },
-  ]
-}
-
-function getLastChangedIndex(groupings: lastChangedGroup[], newest: number, input: number) {
-  const diff = (newest - input)
-  let index = 0;
-  while(index < groupings.length-1) {
-    if (diff < groupings[index+1].epoch) return index
-    else index++
-  }
-  return index
-}
-
-class SpectrumTranslater {
-  readonly scale: number
-  readonly offset: number
-  readonly target_max: number
-  readonly target_min: number
-
-  constructor(input_min: number, input_max: number, target_min: number, target_max: number) {
-    this.scale = (target_max - target_min) / (input_max - input_min)
-    this.offset = input_min * this.scale - target_min
-    this.target_max = target_max
-    this.target_min = target_min
-  }
-
-  translate(input: number) {
-    return input * this.scale - this.offset
-  }
-
-  inverseTranslate(input: number) {
-    return this.target_max - this.translate(input) + this.target_min
-  }
-}
-
-class TruckFactorTranslater {
-  private readonly min_lightness = 50
-  private readonly max_lighness = 90
-  private readonly step: number
-
-  constructor(author_count: number) {
-    this.step = (author_count <= 1) 
-      ? 1 
-      : this.step = (this.max_lighness - this.min_lightness) / Math.floor(Math.log2(author_count))
-  }
-
-  getColor(value: number) {
-    const level = Math.floor(Math.log2(value))
-    return `hsl(0,75%,${this.min_lightness + (level * this.step)}%)`
-  }
-
-  setColor(blob: HydratedGitBlobObject, cache: MetricCache) {
-    cache.colormap.set(blob.path, this.getColor(Object.entries(blob.unionedAuthors?.HISTORICAL ?? []).length))
-  }
-}
-
-class HeatMapTranslater {
-  readonly translater: SpectrumTranslater
-  readonly min_lightness = 50
-  readonly max_lightness = 95
-
-  constructor(min: number, max: number) {
-    this.translater = new SpectrumTranslater(min, max, this.min_lightness, this.max_lightness)
-  }
-
-  getColor(value: number): string {
-    return `hsl(20,100%,${this.translater.inverseTranslate(value)}%)`
-  }
-
-  setColor(blob: HydratedGitBlobObject, cache: MetricCache) {
-    cache.colormap.set(blob.path, this.getColor(blob.noCommits))
   }
 }
