+++ src/metrics/mostCommits.ts
@@ -1,0 +1,21 @@
+import { HydratedGitBlobObject } from "~/analyzer/model"
+import { MetricCache } from "./metrics"
+import { SpectrumTranslater } from "./metricUtils"
+
+export class CommitAmountTranslater {
+    readonly translater: SpectrumTranslater
+    readonly min_lightness = 50
+    readonly max_lightness = 95
+  
+    constructor(min: number, max: number) {
+      this.translater = new SpectrumTranslater(min, max, this.min_lightness, this.max_lightness)
+    }
+  
+    getColor(value: number): string {
+      return `hsl(20,100%,${this.translater.inverseTranslate(value)}%)`
+    }
+  
+    setColor(blob: HydratedGitBlobObject, cache: MetricCache) {
+      cache.colormap.set(blob.path, this.getColor(blob.noCommits))
+    }
+  }
